requirements.txt
absl-py>=2.1.0
addict>=2.4.0
aiofiles>=23.2.1
aiohappyeyeballs>=2.4.0
aiohttp>=3.10.6
aiosignal>=1.3.1
albumentations>=0.5.2
altair>=5.4.1
annotated-types>=0.7.0
antlr4-python3-runtime>=4.8
anyio>=4.5.0
argon2-cffi>=23.1.0
argon2-cffi-bindings>=21.2.0
asttokens>=2.4.1
astunparse>=1.6.3
async-timeout>=4.0.3
attrs>=24.2.0
av>=12.3.0
backcall>=0.2.0
beautifulsoup4>=4.12.3
bleach>=6.1.0
blinker>=1.8.2
braceexpand>=0.1.7
cachetools>=5.5.0
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.3.2
click>=8.1.7
colorama>=0.4.6
contourpy>=1.1.1
cycler>=0.12.1
decorator>=5.1.1
defusedxml>=0.7.1
distro>=1.9.0
docopt>=0.6.2
docstring_parser>=0.16
easydict>=1.9
einops>=0.8.0
exceptiongroup>=1.2.2
executing>=2.1.0
fastapi>=0.115.0
fastjsonschema>=2.20.0
ffmpeg>=1.4
ffmpy>=0.4.0
filelock>=3.16.1
flatbuffers>=24.3.25
fonttools>=4.54.0
frozenlist>=1.4.1
fsspec>=2024.9.0
future>=1.0.0
gast>=0.4.0
gdown>=5.2.0
gitdb>=4.0.11
GitPython>=3.1.43
google-api-core>=2.24.0
google-auth>=2.35.0
google-auth-oauthlib>=1.0.0
google-cloud-aiplatform>=1.76.0
google-cloud-bigquery>=3.27.0
google-cloud-core>=2.4.1
google-cloud-resource-manager>=1.14.0
google-cloud-storage>=2.19.0
google-crc32c>=1.5.0
google-pasta>=0.2.0
google-resumable-media>=2.7.2
googleapis-common-protos>=1.66.0
gradio>=4.44.0
gradio_client>=1.3.0
grpc-google-iam-v1>=0.14.0
grpcio>=1.69.0
grpcio-status>=1.69.0
h11>=0.14.0
h5py>=3.11.0
hickle>=5.0.3
httpcore>=1.0.5
httpx>=0.27.2
huggingface-hub>=0.25.1
hydra-core>=1.1.0
idna>=3.10
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
imgaug>=0.4.0
importlib_metadata>=8.5.0
importlib_resources>=6.4.5
iniconfig>=2.0.0
ipython>=8.12.3
jedi>=0.19.2
Jinja2>=3.1.4
jiter>=0.7.1
joblib>=1.4.2
jsonschema>=4.23.0
jsonschema-specifications>=2023.12.1
jupyter_client>=8.6.3
jupyter_core>=5.7.2
jupyterlab_pygments>=0.3.0
keras>=2.13.1
kiwisolver>=1.4.7
kornia>=0.5.0
lazy_loader>=0.4
libclang>=18.1.1
Markdown>=3.7
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
matplotlib>=3.7.5
matplotlib-inline>=0.1.7
mdurl>=0.1.2
minio>=7.2.10
mistune>=3.0.2
model-index>=0.1.11
mpmath>=1.3.0
multidict>=6.1.0
narwhals>=1.21.1
nbclient>=0.10.0
nbconvert>=7.16.4
nbformat>=5.10.4
networkx>=3.1
numpy>=1.24.3
nvidia-cublas-cu12>=12.1.3.1
nvidia-cuda-cupti-cu12>=12.1.105
nvidia-cuda-nvrtc-cu12>=12.1.105
nvidia-cuda-runtime-cu12>=12.1.105
nvidia-cudnn-cu12>=8.9.2.26
nvidia-cufft-cu12>=11.0.2.54
nvidia-curand-cu12>=10.3.2.106
nvidia-cusolver-cu12>=11.4.5.107
nvidia-cusparse-cu12>=12.1.0.106
nvidia-nccl-cu12>=2.18.1
nvidia-nvjitlink-cu12>=12.6.68
nvidia-nvtx-cu12>=12.1.105
oauthlib>=3.2.2
omegaconf>=2.1.2
openai>=1.54.4
opencv-python>=4.10.0.84
opencv-python-headless>=4.10.0.84
opendatalab>=0.0.10
openmim>=0.3.9
openxlab>=0.0.11
opt-einsum>=3.3.0
ordered-set>=4.1.0
orjson>=3.10.7
packaging>=24.1
pandas>=2.0.3
pandocfilters>=1.5.1
parso>=0.8.4
pexpect>=4.9.0
pickleshare>=0.7.5
pillow>=10.4.0
pipreqs>=0.5.0
pkgutil_resolve_name>=1.3.10
platformdirs>=4.3.6
pluggy>=1.5.0
progressbar2>=4.5.0
prompt_toolkit>=3.0.48
proto-plus>=1.25.0
protobuf>=5.29.3
psutil>=6.0.0
ptyprocess>=0.7.0
pure_eval>=0.2.3
pyarrow>=17.0.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
pycparser>=2.22
pycryptodome>=3.20.0
pydantic>=2.9.2
pydantic_core>=2.23.4
pydeck>=0.9.1
pydub>=0.25.1
pyee>=12.1.1
Pygments>=2.18.0
pyparsing>=3.1.4
PySocks>=1.7.1
pytesseract>=0.3.13
pytest>=8.3.4
python-dateutil>=2.9.0.post0
python-ffmpeg>=2.0.12
python-multipart>=0.0.10
python-utils>=3.8.2
pytorch-lightning>=1.2.9
pytz>=2024.2
PyWavelets>=1.4.1
PyYAML>=6.0.2
pyzmq>=26.2.0
referencing>=0.35.1
requests>=2.32.3
requests-oauthlib>=2.0.0
rich>=13.8.1
rpds-py>=0.20.1
rsa>=4.9
ruff>=0.6.7
safetensors>=0.4.5
scenedetect>=0.6.4
schedule>=1.2.2
scikit-image>=0.17.2
scikit-learn>=0.24.2
scipy>=1.10.1
seaborn>=0.13.2
segment_anything
semantic-version>=2.10.0
shapely>=2.0.6
shellingham>=1.5.4
six>=1.16.0
smmap>=5.0.1
sniffio>=1.3.1
soupsieve>=2.6
stack-data>=0.6.3
starlette>=0.38.6
streamlit>=1.40.1
sympy>=1.13.3
tabulate>=0.9.0
tenacity>=9.0.0
tensorboard>=2.13.0
tensorboard-data-server>=0.7.2
tensorflow>=2.13.1
tensorflow-estimator>=2.13.0
tensorflow-io-gcs-filesystem>=0.34.0
termcolor>=2.4.0
thinplate
threadpoolctl>=3.5.0
tifffile>=2023.7.10
timm>=1.0.9
tinycss2>=1.4.0
toml>=0.10.2
tomli>=2.0.1
tomlkit>=0.12.0
torch>=2.1.2
torchmetrics>=0.2.0
torchvision>=0.16.2
tornado>=6.4.1
tqdm>=4.66.5
traitlets>=5.14.3
triton>=2.1.0
typer>=0.12.5
typing_extensions>=4.12.2
tzdata>=2024.2
urllib3>=2.2.3
uvicorn>=0.30.6
watchdog>=4.0.2
wcwidth>=0.2.13
webdataset>=0.2.100
webencodings>=0.5.1
websockets>=12.0
Werkzeug>=3.0.4
wrapt>=1.16.0
yapf>=0.40.2
yarg>=0.1.9
yarl>=1.12.1
zipp>=3.20.2
